Renamed componets and removed View from their names
