+++ frontend/views/ProjectDetail.vue
